requirements.txt
google-search-results
langchain
langchain-community
langchain-core
langchain-openai
langgraph
openai
python-dateutil
python-dotenv
requests
streamlit

# Security dependencies
pydantic>=2.5.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
structlog>=23.2.0
